+++ app/models/assignment_properties.rb
@@ -57,7 +57,7 @@ class AssignmentProperties < ApplicationRecord
   validates :start_time, presence: { if: :is_timed }
   validate :start_before_due, if: :is_timed
   validate :not_timed_and_scanned
-  validates :remote_autotest_settings_id, uniqueness: { allow_nil: true }
+  validate :remote_autotest_settings_id_check
 
   STARTER_FILE_TYPES = %w[simple sections shuffle group].freeze
 
@@ -125,5 +125,19 @@ class AssignmentProperties < ApplicationRecord
   def not_timed_and_scanned
     msg = I18n.t('activerecord.errors.models.assignment_properties.attributes.is_timed.not_scanned')
     errors.add(:base, msg) if is_timed && scanned_exam
+  end
+
+  # Ensure remote_autotest_settings_id is unique for a given autotester
+  def remote_autotest_settings_id_check
+    unless self.remote_autotest_settings_id.nil?
+      assign_prop = AssignmentProperties.joins(assignment: :course).where(
+        remote_autotest_settings_id: self.remote_autotest_settings_id,
+        course: { autotest_setting_id: self.assignment.course.autotest_setting_id }
+      )
+      if assign_prop.count == 1 && self.id != assign_prop.first.id
+        errors.add(:remote_autotest_settings_id,
+                   'remote_autotest_settings_id has already been taken by the same autotester.')
+      end
+    end
   end
 end