+++ Changelog.md
@@ -13,6 +13,7 @@
 - Ensures mark levels on a rubric criterion are properly scaled when its max mark is updated (#7311)
 - Refactor contributors list in About section to read from markus-contributors.txt (#7374)
 - Optimized the querying of submissions when assigning graders (#7381)
+- Update remote_autotest_settings_id validation to be unique to an autotester (#7393)
 
 ### 🔧 Internal changes